[main] validation finished

+++ src/main/java/com/example/cash_register/controller/commands/expert/DeleteProductCommand.java
@@ -1,9 +1,12 @@
 package com.example.cash_register.controller.commands.expert;
 
 import com.example.cash_register.controller.commands.Command;
+import com.example.cash_register.model.enums.Roles;
 import com.example.cash_register.model.service.ProductService;
 
 import javax.servlet.http.HttpServletRequest;
+
+import static com.example.cash_register.utils.Validator.*;
 
 public class DeleteProductCommand implements Command {
 
@@ -16,11 +19,25 @@ public class DeleteProductCommand implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        Long id = Long.parseLong(request.getParameter("id"));
+
+        if (!isAllowed(request, Roles.COMMODITY_EXPERT)) {
+            return "redirect:/login.jsp";
+        }
+
+        String idStr = request.getParameter("id");
+        String pageStr = request.getParameter("page");
+        String sort = request.getParameter("sort");
+
+        if(!isIdValid(idStr)||!isIdValid(pageStr)||nullOrEmpty(sort)){
+            request.setAttribute("exception", "Wrong data entered");
+            return "/WEB-INF/exception.jsp";
+        }
+
+        Long id = Long.parseLong(idStr);
         productService.deleteProduct(id);
 
-        Integer page = Integer.parseInt(request.getParameter("page"));
-        String sort = request.getParameter("sort");
+        Integer page = Integer.parseInt(pageStr);
+
         request.setAttribute("sort", sort);
         request.setAttribute("pagesCount", productService.getPagesCount());
         request.setAttribute("page", page);
