+++ src/main/java/com/example/cash_register/controller/commands/manager/DeleteProductFromReceipt.java
@@ -2,9 +2,13 @@ package com.example.cash_register.controller.commands.manager;
 
 import com.example.cash_register.controller.commands.Command;
 import com.example.cash_register.controller.exceptions.NoSuchProductException;
+import com.example.cash_register.model.enums.Roles;
 import com.example.cash_register.model.service.ReceiptService;
 
 import javax.servlet.http.HttpServletRequest;
+
+import static com.example.cash_register.utils.Validator.isAllowed;
+import static com.example.cash_register.utils.Validator.isIdValid;
 
 public class DeleteProductFromReceipt implements Command {
 
@@ -16,9 +20,20 @@ public class DeleteProductFromReceipt implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        Long productId = Long.parseLong(request.getParameter("productId"));
-        Long receiptId = Long.parseLong(request.getParameter("receiptId"));
+        if (!isAllowed(request, Roles.CASHIER_MANAGER)) {
+            return "redirect:/login.jsp";
+        }
 
+        String prodIdStr = request.getParameter("productId");
+        String recIdStr = request.getParameter("receiptId");
+
+        if (!isIdValid(prodIdStr) || !isIdValid(recIdStr)) {
+            request.setAttribute("exception", "Wrong id format while deleting");
+            return "/WEB-INF/exception.jsp";
+        }
+
+        Long productId = Long.parseLong(prodIdStr);
+        Long receiptId = Long.parseLong(recIdStr);
 
         receiptService.deleteProductFromReceipt(receiptId, productId);
