+++ src/main/java/com/example/cash_register/controller/commands/manager/DeleteReceiptCommand.java
@@ -1,9 +1,13 @@
 package com.example.cash_register.controller.commands.manager;
 
 import com.example.cash_register.controller.commands.Command;
+import com.example.cash_register.model.enums.Roles;
 import com.example.cash_register.model.service.ReceiptService;
 
 import javax.servlet.http.HttpServletRequest;
+
+import static com.example.cash_register.utils.Validator.isAllowed;
+import static com.example.cash_register.utils.Validator.isIdValid;
 
 public class DeleteReceiptCommand implements Command {
     private final ReceiptService receiptService;
@@ -15,7 +19,17 @@ public class DeleteReceiptCommand implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        Long id = Long.parseLong(request.getParameter("id"));
+        if (!isAllowed(request, Roles.CASHIER_MANAGER)) {
+            return "redirect:/login.jsp";
+        }
+
+        String idStr = request.getParameter("id");
+
+        if (!isIdValid(idStr)) {
+            request.setAttribute("exception", "Wrong id format");
+            return "/WEB-INF/exception.jsp";
+        }
+        Long id = Long.parseLong(idStr);
         receiptService.deleteReceipt(id);
 
         //  Integer page = Integer.parseInt(request.getParameter("page"));
